Remove RescoreBuilder interface, rename AbstractRescoreBuilder to RescoreBuilder
